backend/requirements.txt
blinker==1.6.2
blurhash==1.1.4
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
decorator==5.1.1
Flask==2.3.2
Flask-Cors==4.0.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Mastodon.py==1.8.1
numpy==1.25.1
pandas==2.0.3
psycopg2==2.9.6
pyparsing==3.1.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
